automation/requirements.txt
requests
feedparser
google-genai
python-dotenv
pyyaml
notebooklm-mcp-cli
